python-image/requirements.txt
pymupdf
pandas
pytesseract
langdetect
icecream
pillow
multiprocessing-logging
openpyxl

